Update modules files
