+++ index.html
@@ -72,6 +72,6 @@
     <footer>
       <p class="footer-p">Copyright&copy; 2022, Awesome Books</p>
     </footer>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
