+++ index.js
@@ -1,12 +1,7 @@
-class Book {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-}
-  
+import Book from './modules/book.js';
+import Nav from './modules/nav.js';
 const bookStorage = [];
-  
+
 class CreateBook {
   static addBook(newBook, index) {
     const library = document.querySelector('#books-container');
@@ -61,26 +56,26 @@ class CreateBook {
     });
   }
 }
-  
+
 const form = document.querySelector('.books-form');
-  
+
 form.addEventListener('submit', (e) => {
   e.preventDefault();
   const title = document.querySelector('#title').value;
   const author = document.querySelector('#author').value;
-  
+
   if (title !== '' && author !== '') {
     const newBook = new Book(title, author);
     CreateBook.addBook(newBook);
-  
+
     CreateBook.updateLocalStorage(newBook);
-  
+
     document.querySelector('#title').value = '';
     document.querySelector('#author').value = '';
-  
+
     document.querySelector('#title').focus();
   }
-  
+
   const deleteBtn = document.querySelectorAll('.remove-btn');
   deleteBtn.forEach((btn) => {
     btn.addEventListener('click', (e) => {
@@ -90,36 +85,8 @@ form.addEventListener('submit', (e) => {
     });
   });
 });
-  
+
 CreateBook.getLocalStorage();
-
-// buttons
-const navList1 = document.getElementById('list');
-const navList2 = document.getElementById('add-new');
-const navList3 = document.getElementById('contact');
-
-// sections
-const sectionOne = document.getElementById('all-book');
-const sectionTwo = document.getElementById('add-new-book');
-const sectionThree = document.getElementById('contact-form');
-
-navList1.addEventListener('click', (e) => {
-  e.preventDefault();
-  sectionOne.classList.remove('hide');
-  sectionTwo.classList.add('hide');
-  sectionThree.classList.add('hide');
-});
-
-navList2.addEventListener('click', (e) => {
-  e.preventDefault();
-  sectionOne.classList.add('hide');
-  sectionTwo.classList.remove('hide');
-  sectionThree.classList.add('hide');
-});
-
-navList3.addEventListener('click', (e) => {
-  e.preventDefault();
-  sectionOne.classList.add('hide');
-  sectionTwo.classList.add('hide');
-  sectionThree.classList.remove('hide');
-});
+Nav.navList1();
+Nav.navList2();
+Nav.navList3();
