+++ modules/book.js
@@ -1,8 +1,8 @@
-/* class Book {
+class Book {
   constructor(title, author) {
     this.title = title;
     this.author = author;
   }
 }
 
-export default Book; */
+export default Book;
